Make request_timeout_seconds a whole number of seconds
qdrant-client takes its timeout as an int and rounds a fraction up, so a
fractional value was honored by pymilvus and silently changed for Qdrant.
An int is honored exactly by both, and matches max_retry_interval_seconds
on the embedder and language model configurations.

(cherry picked from commit 5e9f36385b39ff645830ec60f4af4b6f544e2736)

+++ packages/server/src/memmachine_server/common/configuration/database_conf.py
@@ -266,8 +266,8 @@ class QdrantConf(YamlSerializableMixin, ApiKeyMixin):
             "is set to match so all replicas confirm writes."
         ),
     )
-    request_timeout_seconds: float = Field(
-        default=30.0,
+    request_timeout_seconds: int = Field(
+        default=30,
         gt=0,
         description="Seconds a request to Qdrant may take before the client gives up.",
     )
@@ -301,8 +301,8 @@ class MilvusConf(YamlSerializableMixin, WithValueFromEnv):
             "Supported values: Strong, Session, Bounded, Eventually."
         ),
     )
-    request_timeout_seconds: float = Field(
-        default=30.0,
+    request_timeout_seconds: int = Field(
+        default=30,
         gt=0,
         description="Seconds a request to Milvus may take before the client gives up.",
     )
